refact llamacpp (merge request !145)
Squash merge branch 'refact-llamacpp-2' into 'master'
1. fixed long prompt hang
2. reduce buffer memory
3. remove magic code of prefix cache config
4. add integration test for interact prefix and non-prefix cache

+++ src/ksana_llm/batch_manager/batch_scheduler/batch_scheduler.cpp
@@ -629,7 +629,7 @@ void BatchScheduler::ScheduleWaiting(size_t &step_token_num_sum, size_t &curr_bl
       NLLM_LOG_DEBUG << "waiting req " << req->req_id << " launch.";
       size_t total_block_num = waiting_total_block_num_list_[visit_idx];
       if (total_block_num > 0) {
-        for (size_t i = 0; i < context_->GetTensorParallelSize(); ++i) {
+        for (int i = 0; i < context_->GetTensorParallelSize(); ++i) {
           std::vector<int> blocks;
           GetBlockManager()->SetDeviceId(i);
           GetBlockManager()->AllocateBlocks(total_block_num, blocks);
